Simplify extract_ocr_predictions function
