+++ robotoff/prediction/ocr/__init__.py
@@ -1,3 +1,3 @@
 # flake8: noqa
-from .core import extract_predictions, get_barcode_from_path, ocr_iter
+from .core import extract_predictions, ocr_iter
 from .dataclass import OCRResult
